small changes to match I/O documentation
- added comments to clarify a few variables
- changed nr_alighting_doors and nr_boarding_doors to ints
- changed naming of o_bus*.dat output files to o_transit*.dat
(Note: the ordering of the output files now differes from the ordering
in the I/O doc, to change this to be consistant one would have to
carefully go through the whole write_busstop_output function)

+++ branches/busmezzo/mezzo_lib/src/busline.h
@@ -763,8 +763,8 @@ public:
 		double boarding_coefficient_, 
 		double alighting_cofficient_, 
 		double dwell_std_error_, 
-		double number_boarding_doors_, 
-		double number_alighting_doors_,
+		int number_boarding_doors_, 
+		int number_alighting_doors_,
 		double share_alighting_front_door_, 
 		double crowdedness_binary_factor_, 
 		double bay_coefficient_, 
@@ -813,8 +813,8 @@ public:
 	double dwell_std_error;
 
 	// only for TCRP functions
-	double number_boarding_doors;
-	double number_alighting_doors;
+	int number_boarding_doors;
+	int number_alighting_doors;
 	double share_alighting_front_door;
 	double crowdedness_binary_factor;
